+++ lib/Cmts.ml
@@ -302,29 +302,27 @@ let init fragment ~debug source asts comments_n_docstrings =
     match locs with
     | [] -> add_cmts t `After ~prev:Location.none Location.none cmts
     | _ -> place t loc_tree locs cmts ) ;
-  let () =
-    let relocate_loc_stack loc stack =
-      List.iter stack ~f:(fun src -> relocate t ~src ~before:loc ~after:loc)
-    in
-    let expr (m : Ast_mapper.mapper) x =
-      relocate_loc_stack x.pexp_loc x.pexp_loc_stack ;
-      Ast_mapper.default_mapper.expr m x
-    in
-    let typ (m : Ast_mapper.mapper) x =
-      relocate_loc_stack x.ptyp_loc x.ptyp_loc_stack ;
-      Ast_mapper.default_mapper.typ m x
-    in
-    let pat (m : Ast_mapper.mapper) x =
-      relocate_loc_stack x.ppat_loc x.ppat_loc_stack ;
-      Ast_mapper.default_mapper.pat m x
-    in
-    let _ =
-      Mapper.map_ast fragment
-        Ast_mapper.{default_mapper with pat; typ; expr}
-        asts
-    in
-    ()
+  let relocate_loc_stack loc stack =
+    List.iter stack ~f:(fun src -> relocate t ~src ~before:loc ~after:loc)
   in
+  let mapper =
+    object
+      inherit Ppxlib.Ast_traverse.map as super
+
+      method! pattern x =
+        relocate_loc_stack x.ppat_loc x.ppat_loc_stack ;
+        super#pattern x
+
+      method! core_type x =
+        relocate_loc_stack x.ptyp_loc x.ptyp_loc_stack ;
+        super#core_type x
+
+      method! expression x =
+        relocate_loc_stack x.pexp_loc x.pexp_loc_stack ;
+        super#expression x
+    end
+  in
+  let _ = Mapper.map_ast fragment mapper asts in
   t
 
 let preserve fmt_x t =