+++ CHANGES.md
@@ -51,6 +51,10 @@
 
 #### New features
 
+#### Internal
+
+  + Use ppxlib instead of ocaml-migrate-parsetree 1.x. (#1482, @emillon)
+
 ### 0.15.0 (2020-08-06)
 
 #### Changes